+++ src/components/index.js
@@ -4,6 +4,7 @@ export { default as Navbar } from './Navbar.jsx';
 export { default as Footer } from './Footer.jsx';
 export { default as Main } from './Main.jsx';
 export { default as AllTimelines } from './AllTimelines';
+export { default as AllTimelines2 } from './AllTimelines2';
 export { default as EditTimeline } from './EditTimeline.jsx';
 export { default as Login } from './Login';
 export { default as SignUp } from './SignUp';
